rearranged routes and started organizing
