--- handler/check_data.go
@@ -1,1 +1,0 @@
-package handler
