+++ routes/api.go
@@ -1,0 +1,26 @@
+package routes
+
+func (r *Router) n() {
+
+}
+func (r *Router) n() {
+
+}
+func (r *Router) n() {
+
+}
+func (r *Router) n() {
+
+}
+func (r *Router) n() {
+
+}
+func (r *Router) n() {
+
+}
+func (r *Router) n() {
+
+}
+func (r *Router) n() {
+
+}
